+++ CHANGES.rst
@@ -5,7 +5,18 @@ Changelog of trs
 1.4 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Added extra column 'booked this year' in PersonView. Handy.
+
+- Added overview of project leaders and managers.
+
+- Nicer feedback in title on selected filters (for persons and projects page).
+
+- Added groups. Including using filters on groups in person and projects views
+
+- Added extra column 'booked this year' in PersonView. Handy.
+
+- Better change overview page, including option to see all projects' changes
+  when you're a manager.
 
 
 1.3 (2014-01-21)